chore: rename to test-utilities folder
