+++ react-components/tests/tests-utilities/fixtures/cadModel.ts
