+++ react-components/tests/tests-utilities/fixtures/cameraManager.ts
