+++ react-components/tests/tests-utilities/fixtures/image360.ts
