+++ react-components/tests/tests-utilities/fixtures/pointCloud.ts
